Update bash aliases for git commit
Before I used to have "gcom" as an alias for "git commit -S -m".
Whereas now I use that for "git commit -S", which opens the git editor
(Vim in my case) for drafting the commit message.  The previous command
is preserved in the alias "gcomm".

For the sake of completeness, setting the git editor is done with this
command:

    git config --global core.editor vim