Fix script for laptop's dual monitor setup
My ThinkPad X220 is often connected to an external monitor via the VGA
port.  The monitor is physically placed above the laptop's screen.  The
original script would set the external display as the primary one, on
the assumption that it made sense to start counting from the top to the
bottom.

This had the side-effect of failing to pass some brightness-related
XRandR options to the laptop's screen.  Though it would still configure
properly the coordinates of each monitor.

With this commit, the LVDS monitor is considered the primary one, fixing
the problem of setting the appropriate brightness and gamma settings.