Integrate dotsmenu
Rewrite the dotsmenu script.  The original concept of quickly accessing
dotfiles remains the same, but the code is written anew.

A key chord is now added to sxhkdrc to invoke dotsmenu directly.

If the script is run from within a terminal, it will open the selected
file in the $EDITOR.  Otherwise it will resort to using (in sequence):
$VISUAL, gvim, xdg-open.