Update sxhkdrc
Add a new key chord chain pattern for node splits:

	super + s ; ...

This is used to create new key bindings for changing the split ratio:

	super + s ; {2,3}
		... {0.50,0.66}

The numbers are a reference to the area of the active node that is being
allocated to the new window: 0.50 = 1/2, 0.66 = 1/3.

A couple of older key chords for balancing/equalising node splits have
switched from the pattern super + a to the one mentioned above.