Major update to the Tempus themes files
This syncs the Tempus themes with the latest changes that were
introduced in Version 0.24.0.20190324 of the Tempus Themes Generator.
See the repo: https://gitlab.com/protesilaos/tempus-themes-generator

The gist of this review is to improve the distinctiveness of the colours
used in certain themes.  A common problem was the distinction between
green and cyan, bright red and bright yellow, blue and cyan, magenta and
bright magenta.  The affected themes should, nonetheless, feel the same
practically the same.