Review GEV code
There were some _potential_ problems with the original design.  The
major one concerns the implementation of `find`.  The results of this
command should not be used in a loop, as it will create problems with
file names taht contain white space or glob characters.

To help me with this review, I used the `shellcheck` utility from the
Debian Buster/Sid archives:

	apt install shellcheck