Change font size for my generic `dmenu` typeface
This file is sourced from various scripts of mine, typically `dmenu`
implementations, to define the font family that is to be used.  The font
size has been changed to better complement the new default sans-serif
typeface.[0]  This specific point size makes the height of a vertical
`dmenu` equal (at least virtually so) to the height of my custom
`lemonbar`, which is the kind of visual consistency I strive for.

[0]: See commit 83249ab6b6838409668833dde139d4e5c936ec9b