Major changes: own_script_update_environment_theme
Review practically everything.  Convert space indentation into tabs.
Rework functions that had lots of repetitive logic and `sed`
invocations.  Improve documentation.

All this is considered a "major update" because it is still pending
review.  A "full review" means that the script is read from top to
bottom and is inspected for any irregular behaviour.