MAJOR UPDATE: Refactor melonpanel
Reworked the entire script.  Simplified and/or improved conditional
logic for the various modules.  Also reworked the colours and styles.

Concerning the available modules, most of those that were previously put
in a separate file and treated as deprecated are now reincorporated in
to the panel.  This is thanks to the improved code structure.