BREAKING CHANGE: remove all custom fonts
Deleted everything related to custom fonts.  These are not "dotfiles"
per se, nor do they add any kind of essential functionality to the
environment.  Besides, the scope of their customisations is narrow
enough to not merit the added trouble of maintaining and distributing
them.

That granted, I strive to remove all external dependencies from my
dotfiles, while relying exclusively on the Debian repos (specifically
those that are enabled for the upcoming release of Debian 10 'buster').

Here are the package names of some excellent monospaced fonts, which are
designed for coding:

1. xfonts-terminus (my new default)
2. fonts-monoid
3. fonts-inconsolata
4. fonts-hack
5. fonts-firacode

Plus the DejaVu family that offers the sublime "DejaVu Sans Mono".