Important tweak to the script that changes themes
The script would check for the path to the dotfiles, using a git
command.  This works fine when running it from _within_ the dotfiles'
directory, otherwise it throws an error.  Tweaking the logic to first
check for a hard-wired path, seems to do the trick for now.

HOWEVER, this snipper (as well as the rest of the script) are in need of
a thorough review, in the interest of code correctness.