BREAKING: BSPWM programmatic colour definitions
A long while ago, I had a script that would run some `bspc config`
commands to set the WM colours.  The reason was that `source`ing the
colours and applying them directly in the `bspwmrc` proved unreliable.

Now the colours are hard-coded in the config file and are controlled by
`own_script_update_environment_theme` and/or its `tempusmenu` interface.

As such, the `own_script_bspwm_config_post_launch` and the relevant
snippet in the `bspwmrc` are removed.