BREAKTHROUGH (!?): script to update running Vim
Add new script that uses `tmux` to find running instances of `vim` and
then send a ":source ~/.vimrc" command to them.

This opens up the possibility to "live reload" all running instances of
the text editor when updating the environment theme.

For context, I want to change Vim's colour scheme via a script and then
somehow tell all open buffers to reload the vimrc in order to get the
new theme.  I tried using Vim's `autocmd` but could not achieve what
I was looking for: the point is to update things from the shell.

NOTE this script still needs some conditional logic and documentation.