Live reload Vim instances!
Add the necessary support for the new script that handles the live
reloading of theme changes to Vim.

This has been the only major problem in my "environment theme switch":
all running instances of `vim` had to be closed and opened again to
inherit the new colours, though only when transitioning from a light to
a dark theme (and vice versa).

For the new script see commit: 8ce68a6512ae0251c661b0489ed6140e5305e97e.