Remove Terminus from "Monospace" alias
This is important to fix the rendering of monospaced fonts by
applications that set an arbitrary point size and/or do not respect
system-wide aliasing rules.  Web browsers are the greatest offenders
in large part because many websites are poorly designed.

Terminus is a fixed-size (bitmap) font that does not scale well.  It is
better to explicitly define that font as the default in text editors or
other contexts where we have full control over its presentation (since
Terminus _is_ a superb typeface).