Further refinements to melonpanel fonts/looks
Increase the point sizes for both Terminus and the fallback "fixed".
Make the panel height 4 pixels larger to better complement the new
fonts.

This is particularly important for using the panel with light themes.
Otherwise a certain "halo effect" is created around the letter forms,
which distorts the text and makes colours harder to tell apart.