Tweak xrandr command for laptop colours
Slightly modify the combination of brightness and gamma values for the
r:g:b space.  This improves the presentation of colours on the screen
and is particularly noticeable when working with light themes: shades of
white are easier to tell apart.

Note that this specific to my hardware and a check exists to make sure
it is mostly likely not applied elsewhere.