PTP: new script to Print Tempus Palette
ptp will output a table with the colour values of the active Tempus
theme.	Each entry inherits the colour it references, by using ANSI
escape sequences.  Colours are presented in their original hexadecimal
format.

ptp will adjust its output depending on the length of the running
console, with the breaking point defined at 80 columns (which seems to
be the default?).