Refinements to melonpanel fonts
Improved the proportions for Terminus so that letter forms with
descenders do not appear squeezed (letter like g and y).

While the new Terminus definition is nominally a 16pt size, it retains
the same overall width as the previous one, which was 14pt.  I _guess_
this is governed by the "80" value in these:

- melonpanel_font='-xos4-terminus-medium-r-normal--14-140-72-72-c-80-iso10646-1'
+ melonpanel_font='-xos4-terminus-medium-r-normal--16-160-72-72-c-80-iso10646-1'

In the same spirit, the fallback font was also updated.  The new
definition is a nominal 18 point size, however it is just an improved
version of the 15pt version.

The changes to the fallback font make it more legible at the cost of
increasing its overall width, to the point where it is slightly wider
than Terminus.