Add Compton toggle
This provides for an easy way to toggle the display compositor with
a key chord chain.

I personally need to disable compton when recording a screen cast,
because (a) my hardware has trouble coping with it, and (b) it creates
screen flickering.

Otherwise, I like having compton enabled, for its shadows and effects.