Tweak tmux status line
No longer use colour6 as the background of the active window.  It just
calls too much attention to itself.  Instead keep colours constant, but
rely on typography.  Now the active window is enclosed in square
brackets and drawn in a bold font.

Given this opportunity, I also tested a few configurations regarding the
response of the status line's right module to the press of the prefix
key.  It is designed to turn its background into colour3, to signify
that the prefix is active and waiting for a key press.  This has
remained unchanged, to denote the special nature of the tmux prefix.