Tweak melonpanel's bspwm indicators
The previous design from commit 2f8756716494b3dd247aa5118b6bd6cd07bc2582
had a total length of 4 characters for each workspace.  This meant that
the name could never be in the very centre, creating excessive spacing
between desktop names.  An odd number is needed.  5 chars would be too
long, so 3 it is.

The new styles are thus:

    Focused desktop ................................ "[NAME]"
    Unfocused desktop .............................. " NAME "
    Urgent desktop ................................. " NAME#"
    Occupied unfocused desktop ..................... " NAME^"
    Active desktop on unfocused monitor ............ " NAME-"

Furthermore, the node flags have an asterisk as their default indicator.
Previously this was the mid dot---not very visible.