BREAKING CHANGE: Disable dunst shortcuts
They would conflict with some of my SXHKD bindings.  A notification
daemon should expose its functionality to a client that can be
controlled via the console.  This fact alone is reasone for me to search
for alternatives, including the possibility of switching to the generic
tool that ships with the MATE desktop (my fallback DE).