Consolidate desktop navigation
The defined key chords now range from 0 through 9.  This replaces the
old approach which was {1-9,10}.

The reason is that desktop 10 can be ambiguous under certain
circumstances, as it might point to desktop 1 instead.  Better keep
everything as a single digit.

Also note that the key chord that changes the active desktop does not
prepend a caret sign to the desktop number.  This used to work before,
so I am assuming the adjustment has to do with the names that
bspwm_dynamic_desktops generates.

The key chord for sending and following a node to a given desktop has
been removed.  This was super + ctrl + shift + {0-9}.  It is an awkward
key combination which does not offer a compelling feature while also
posing a possible source of confusion with other more useful key chords
(such as super + shift + {0-9} or super + ctrl + {0-9}).