Tweak Xterm blink settings
This makes the blink's on-off transition less abrupt.

Recall that we use blinking to circumvent the fact that Xterm does not
properly inherit/adapt the background colour (say red) under the cursor
while editing text in Vim.  This can result in indecipherable text and
difficulties with identifying the cursor's point visually.