Update melonpanel documentation for lemonbar-xft
I have a fork of lemonbar that supports Xft (the fontconfig library).
This commit just documents all the relevant information, in case any
user is interested to OPT IN.

I have decided to keep as default the configurations that work with the
Debian/upstream package.  I do not want to force anyone to compile stuff
from source.