MINOR REVIEW: continuous input for node resize
	super + c : {h,j,k,l}

Pressing `super + c` enters the mode.  Then the tail of the key chord,
the part after the colon sign, can be used on its own to trigger the
action: expand or contract the focused node in the given direction (Vim
keys).  The Escape key exits the mode.

This REPLACES the previous mapping for this key chord, which would
switch focus in the given direction.  I believe that was not
a particularly useful implementation.