POTENTIALLY GREAT: xterm cursor reverse colour
I discovered this by accident.  Serendipity, if you will.  What it does
is to tell Xterm to paint the cursor in the background colour of the
character under it.  So e.g. a blue letter turns the cursor block blue.

This is desirable because there are many cases, such as in Vim or the
Tmux copy-mode, where the cursor becomes practically invisible.

My tests so far show that this problem is now fixed.