Tweak multihead settings
Adjust the way monitors are placed along the 2d coordinates.  This is to
accomodate the default behaviour of treating the display that starts at
0x0 as the "primary" one.

To this end, assign desktop "1" to the LVDS1 monitor (laptop display
starting at 0x0) and desktop "8" to the VGA1 monitor (which is to the
right of the laptop).