BREAKING CHANGES: control session with Hyper key
My entire session is no longer dependent on using the Super key (mod4).
Its function is now performed by Hyper (mod3).  This new modifier key is
defined in `.Xmodmap', which is included in the same Stow package as the
core BSPWM/SXHKD configs (because it is equally important to them).

The `.Xmodmap' enables Hyper and binds it to the Caps lock and the right
"Windows" key.  This means that there no longer exists a toggle for
writing all capital letters.  The Super key is also limited to just the
left Windows key.  This is a tentative design that might be reviewed
upon further consideration.

* * *

The present commit is merely _the first step_ in a major review of my
custom desktop session.  The overarching objective is to eventually
merge my bespoke BSPWM setup with my newer Emacs environment.

For the time being, my "dotemacs" is available from my website[0] and is
stored in a separate repo,[1] but will ultimately be merged with my
dotfiles once all the prerequisites are met.

Expect far-reaching changes to practically everything that I now do
inside of Emacs (writing, emailing, reading…) and is still covered by
specialised tools (vim, mutt, newsboat, ncmpcpp…).  For more on that,
see my video blog on the Emacs mindset and the Unix philosophy (ignore
my remarks on Xfce, as I did not know about `mod3' back then).[2]

[0]: https://protesilaos.com/dotemacs
[1]: https://gitlab.com/protesilaos/dotemacs
[2]: https://protesilaos.com/codelog/2019-08-09-vlog-emacs-unix/

* * *

If anyone reviewing this has read my "Prot's Dots For Debian",[4] they
should be reminded that PDFD does not track this repo directly.  It has
its own "snapshot" of my dotfiles.[5]

I will not be pushing any of these breaking changes to PDFD for the time
being.  That is a longer term objective that _might happen_ with the
release of Debian 11 'bullseye' (about 22 months from now).

[4]: https://protesilaos.com/pdfd
[5]: https://gitlab.com/protesilaos/cpdfd