BREAKING: Control Emacs Bongo
This uses the `emacsclient -e' to evaluate the functions that control my
current music player.

What is great about this otherwise minor addition is the possibilities
it opens up: I now know how to control Emacs remotely!

With regard to my outlook going forward, I will be removing all the
MPD-related configs.  I have realised that my music setup does not need
to be as elaborate.