Minor refinements to live theme reload
The running settings daemon allows us to update the GTK themes of all
running applications.  For me this is important because, for example, I
might have a web browser open with a few tabs and another graphical
application on the side, which I do not want to close in order to apply
the new styles.

It seems that the issue I experienced with the de-activation of Xmodmap
was due to a specific application and only while trying to switch
keyboard layouts.  Will need to investigate this further and submit a
bug report for that application.