FIX (?) tempusmenu: use `rofi -dmenu -only-match'
The previous setup would not interpret a cancellation/exit/escape as the
termination of `rofi'.  Instead it would evaluate it as valid input.  It
would thus corrupt configuration files by replacing the colour values
with empty strings.

The `-only-match' demands valid input, but has a downside of disabling
all methods of quitting the prompt.  I am sure about said downside.
Still need to investigate this.