BREAKING UPGRADE: full review of multihead scripts
Prior to this major cleanup I had a rather convoluted logic for
determining settings based on whether I was on my laptop or not and
whether an external monitor was connected.

While it worked, it was not elegant.

Now one script defines the XrandR settings and another sets the desktops
accordingly.  These are loaded by bspwmrc before anything else happens.