Remove settings daemon dependency. Use xsettingsd.
The reason I was using a settings daemon was to be able to live reload
GTK applications while performing a theme switch (via `tempus').

But those daemons implements lots of extra configurations that break my
setup in subtle ways.  For example, the Xfce one (`xfsettingsd'), has a
seemingly unmodifiable keybinding for Super+p.  This breaks an Emacs
binding that I have.  Unacceptable!