Emacs modus themes: remove `dired-subtree' bg
Setting background values for this package's various depth levels breaks
the consistency of established metaphors.  The line of a Dired mark or
flag is not coloured as intended because dired-subtree's styles take
precedence.

Unless there is a way to change the prioritisation of faces,
`dired-subtree-depth-*-face' must have their `:background' set to `nil'.