Emacs: update fonts.  Revert to my patched "Hack".
Also fixes an issue with the documentation of `prot/desktop-fonts',
which was brought up by GitLab user Fourchaux.[0] Thanks for that!

By the by, I plan to write a clean shell script that can be used to
reproduce my patched version of Hack.[1] I did not do it the first time
round, because I was testing things and then had to address some more
pressing issues.

[0]: https://gitlab.com/protesilaos/dotfiles/issues/1
[1]: https://gitlab.com/protesilaos/hack-font-mod