Emacs modus themes: refinements to Ivy and Swiper
Ivy can now highlight different matching groups of a regexp.  This makes
it easier to evaluate the accuracy of a query.

Swiper does the same in the minibuffer, but not in the actual window.
For Swiper's matching line in the window, the current match now appears
as a uniformly styled line.  This is meant to avoid the issue of
creating a rainbow effect throughout the frame, while also helping
distinguish the window's results from those of the minubuffer.

See: https://gitlab.com/protesilaos/modus-themes