Emacs: REVIEW Ivy and Prescient
Yesterday I pushed some changes that outsourced all parameters for
`ivy-re-builders-alist' to `prescient'.  I am reverting that set of
changes because it totally messes up with the highlighting that Ivy
implements.  I find that colours are very important in this context, so
I prefer to remain aligned with the default approach.

Documentation has been updated to account for these adjustments.