STOW: desktop/menu/rofi entry for custom Emacs
This is useful for the binary that is produced by compiling Emacs from
source.  It makes Emacs accessible via various menu interfaces,
including the option that `rofi' has to only show executables with a
desktop entry.

There is an option to specify a path in the `*.desktop' file, but it is
not required if you already have the binary in your PATH.  In my case,
the binary is in `~/bin', which is appended to the PATH (the binary
itself is gitignored).