Emacs: UPDATE ALL FONT SETTINGS (revert to "Hack")
This is in response to feedback I received for some of my recent screen
casts.  No matter the size I used, Iosevka (my previous choice) always
felt small to some people.  My guess is this is due to the compressed
nature of that typeface: even Iosevka's "extended" version is somewhat
narrow when compared to the spacious/comfortable Hack font (and all
others in the DejaVu Sans Mono lineage).