Emacs: simplify fonts. Just use "Hack"
Will not be using Iosevka at all.  The change of font families between
the laptop and desktop contexts disorients me.

This concludes a ~1-month long experiment of switching to Iosevka as my
main font.  Even the "extended" version of that typeface feels less
readable than Hack, something that multiple viewers of my videos
commented on.