Emacs: MAJOR REVIEW WITH BREAKING CHANGES FOR IVY
I am deprecating Ivy and all the packages there were related to it in
favour of the built-in `icomplete' tool, enhanced by my own functions.

Ivy is great but I need to do this in order to grow as an Emacs user and
to expand my knowledge of elisp.  A more minimal tool allows me to solve
problems myself, thus providing the impetus to further research and
experimentation.

I understand that many users do not share the same motives and would
rather use something that just works well.  This is why I am preserving
all my configurations and extensions for Ivy (and the others).  A
prominent deprecation notice is added to every relevant section of this
document.