Emacs: banish whimsical indentation
This all started when I wanted to change the level of a heading and it
would then require from me to make indentation adjustments.  Never liked
it.  Never will.

Let us keep it simple on the visual side, since the typographic elements
combined with the theme's colours are already enough to figure out what
is going on.