Emacs USLS: fix several issues with corrupted vars
+ The categories' history would get corrupted by receiving a list from
the 'completing-read-multiple' prompt.

+ Because the categories' list could often be a single element long,
other problems could arise, such as 'mapcar' throwing an error.

+ History-related 'defvar' have their value set to an empty list, for
the sake of clarity (though I am not sure this is makes any difference).

+ With 'usls--directory' we make sure that we do not create malformed
filesystem paths when catenating 'usls-directory' with some other
string.

+ Corrected the 'usls-id' docstring.

+ All the 'usls--categories-*' functions are complementary to the
command 'usls-new-note'.  They make it easier to understand how things
are pieced together.  This is the path forward for other subroutines of
that or other commands.