prot-search: fix isearch count for repeat searches
This concerns prot-search-isearch-repeat-forward which I bind to the
down arrow key.  Because of the backward motion to keep point at the
beginning of the match, it would reset the counter that isearch shows.
Looking at it again, placing the point at the end of the match while
moving downward feels natural (and at the beginning while moving
upward).