prot-common: define macro for manual packages
This is part of the solution to the problems reported in issue 24:
<https://gitlab.com/protesilaos/dotfiles/-/issues/24>.

Fundamentally, my config is not reproducible in its current state
because it already assumes the presence of all those packages and does
not try to either install them or at least fail gracefully.

I prefer not to auto-install anything: the user must understand what is
going on and remain in control of the process.  So this macro will just
log an error but otherwise skip the execution of a package's code.