init: add EXPERIMENTAL auto-install package setup
Users will need to write a basic-init.el file and place it in the same
directory as my init.el and prot-emacs.org.

The basic-init.el is supposed to contain only "basic" configurations
that control how the rest of my setup will behave.  The only variable
that currently matters in that regard is `prot-emacs-autoinstall-elpa'.

So set that variable to 't' in the basic-init.el and all packages that
are listed on GNU ELPA or MELPA will be installed automatically when the
main configuration file is loaded.  This still does not cover the
handful of packages I handle manually.

This commit continues on the work being discussed in issue 24:
<https://gitlab.com/protesilaos/dotfiles/-/issues/24>.