prot-emacs: use 'cat' as pager in Emacs
This is useful when calling commands that normally pass their output to
the pager.  Because Eshell is not a terminal emulator, it fails to
perform well in the typical scenario of using 'less'.  That is not a
problem, however: using 'cat' just means that we can scroll through the
buffer the same way we do everywhere in Emacs.

I only noticed this issue recently when I was trying various commands
for prot-vc.el and the like.