Revert commit ef2f8cd
I made a mistake last night: the function I had was not a predicate, but
a convenient way to write negative numbers for cases where the user
passes a numeric argument (e.g. C-u 5 and a backward motion should be
internally read as -5, but the user does not need to worry about adding
the negation because the backward motion implies that).