prot-simple: fix line duplication on mark change
Here is the bug:

Say you are in the middle of a buffer and go back to the top with M-<.
Then you try to duplicate the line at point by passing a prefix argument
to 'prot-simple-copy-line-or-region'.  The newly copied line is placed
in the middle of the buffer where the mark was last set (because all
movement motions set the mark).