prot-eww: make feed transform robust
The previous design would fail to deliver the right URL when (1) the
value of (plist-get eww-data :url) did not correspond to the root of the
domain, and (2) the feed had an absolute URL that extended the root
domain.

So instead of testing for the presence of the current URL, we just check
whether any URL is on display and act only if it is not.