prot-minibuffer: use the default "*Completions*"
+ Copy the relevant code for completion cycling from prot-embark.el to
  prot-minibuffer.el.  Tweak it where necessary.

+ Define a method to hide shadowed file name parts.

+ Add a motion for moving backward in file system paths.

+ Update all documentation and settings to work in the new setup.  Also
  make miscellaneous refinements.

For the cycling part, I have been meaning to make this happen for a
while now.  In particular, ever since I discovered that Emacs28 can show
the "*Completions*" as a single column.  I had also realised that my
extras for completion cycling in prot-embark.el had no real dependency
on Embark.

The interaction model should be largely identical to the one before.
However, I still need to make Embark's collection snapshots capture
Marginalia's annotations.  This worked in the status quo ex ante but now
only gets the completion candidates.

As for file name shadowing and the backward motion in file paths, those
are quality-of-life improvements that I used to enjoy when I had
Icomplete, but which I had forgotten about in the months since I
switched to the default minibuffer.  This refactoring gave me the
opportunity to bring them back.