prot-minibuffer: disable workaround for snapshots
I just updated to the latest version of Embark and everything seems to
work now.

To recap, the problem was that 'embark-collect-snapshot' would not store
annotations produced by Marginalia, whereas my simplistic workaround
would do the trick.  However, it was never as good as Embark because it
would not retain the default action of the completion candidates.