prot-minibuffer: simplify relative faces
The old design was necessary when it was still part of
prot-embark.el (and was a bit different to account for Embark's variant
of live updating completions).  Before we wanted to remove my
live-completions-specific styling in snapshots of Embark collect buffers
but otherwise keep it for the completion process.  Whereas now those
tasks are handled in two distinct buffers (the built-in Completions' and
Embarks').