prot-simple: fix new line above when narrowed
The previous design would fail when the buffer was narrowed, because
point 1 would likely be out of range: 'line-number-at-pos' accepts and
ABSOLUTE argument, but this did not seem to work in my case.

Narrowed buffers are fairly common in my workflow.  They also appear
without any manual interventions, such as in message composition
buffers.