Rewrite `prot-eww-{jump-to,visit}-url-on-page in prot-eww.el
From 92603469fbc3eda2d0971c7c72b8b362e8995431 Mon Sep 17 00:00:00 2001
From: Abhiseck Paira <abhiseckpaira@disroot.org>
Subject: [PATCH 1/2] Rewrite `prot-eww-{jump-to,visit}-url-on-page in
 prot-eww.el

The functions `prot-eww-jump-to-url-on-page'[1] and
`prot-eww-visit-url-on-page'[2] are broken into 3 parts. The 3rd part is
the auxiliary function `prot-eww--capture-url-on-page'[3]. [3] abstracts
the common pattern in [1] and [2] which is to get all the links on a web
page and returning a list of strings with those links in a certain way.