prot-eww: add prot-eww-browse-dwim INITIAL-INPUT
It was removed in commit db95a1e, but I find there is a difference in
usability.  With the INITIAL-INPUT present, we can easily edit the
current URL.  For example, I am here:

    https://protesilaos.com/contact

I invoke prot-eww-browse-dwim and call backward-kill-word to delete the
"contact" part, then RET to go to the home page.  This is faster than
using M-n (although I admit it is a minor issue).