prot-eww: introduce prot-eww-quit command and more
From 0493d766f6990262edc3ab2fa5430e0c3479a346 Mon Sep 17 00:00:00 2001
From: Abhiseck Paira <abhiseckpaira@disroot.org>
Subject: [PATCH] prot-eww: introduce prot-eww-quit command and more

A new command called `prot-eww-quit' has been added. It does 3 three
things when called:

1. Delete all the cookies. This behavior can be controlled by the
variable `prot-eww-delete-cookies'. By default it's set to non-nil means
it deletes all the cookies. This is helpful if you want to login to
services with eww, but also want your cookies to be deleted between
browsing sessions.

2. Save the browsing history (Turned off by default). Two variables are
exposed to the user to customize various things about it:

  a. `prot-eww-save-history-place': Where to save history file. By default
  it saves the history file in `user-emacs-directory' by the name
  "prot-eww-visited-history".

  b. `prot-eww-save-visited-history': By default it's set to nil that
  means browsing history will not be remembered. If you want browsing
  history to be remembered and loaded set it to non-nil.

3. Kill all eww buffers. The last thing prot-eww-quit command does is to
kill all the eww buffers. If your (current-buffer) is an eww buffer,
it'll kill that too.

`prot-eww-quit' also asks for confirmation when you call it. It presents
a y-or-n prompt if `prot-eww-save-visited-history' is non-nil. This is
because if you give a wrong answer you don't lose much as your history
is being saved anyways. But if `prot-eww-save-visited-history' is
nil (default) it will give you a yes-or-no prompt.