prot-minibuffer: fix nil completion category bug
I found the bug after trying to select an entry from the Completions'
buffer and would get an error that could ultimately be traced back to
'completion--field-metadata'.  This would happen with regular completion
commands, such as 'switch-to-buffer', but also with some custom ones,
such as 'prot-bongo-playlist-insert-playlist-file' which also has a
'completing-read-multiple' prompt.

This used to work in the past, but no longer did in my last two builds
from emacs.git trunk.  I still do not know why, though this patch seems
to fix things for good.