feat(prot-eww): jump-to-url-on-page works on visible portion
From 46b3e7369fec91a2843040906157451cec9ce2cc Mon Sep 17 00:00:00 2001
From: Abhiseck Paira <abhiseckpaira@disroot.org>
Subject: [PATCH 2/2] feat(prot-eww): jump-to-url-on-page works on visible
 portion

By default prot-eww-jump-to-url-on-page now works on the visible portion
of the buffer in a window. If prefix argument is given then works on the
whole web page (buffer).

A macro has been defined with the name prot-eww-act-visible-window which
replaces a function with the same name. This macro runs BODY forms
inside narrowed region. This narrowed region is the visible portion of
the buffer in a window. This macro returns the last value BODY forms.