prot-tab: implement tmux-like status line
I have known about this for several weeks (months?) now and I would use
it occasionally when I had multiple tabs open.  But I had never actually
checked the source code of tab-bar.el.

It turns out that it is fairly easy to move all the relevant mode line
elements up there.  Maybe there are a few that are missing, so consider
this a work-in-progress.