bspwm: BREAKING replace lemonbar with polybar
This is because polybar lets me set different intervals for individual
components, whereas my lemonbar script was inefficient as it would poll
every second.  I tried to figure out how to do it, but could not make it
work.

I am building polybar using a custom PKGBUILD which removes the i3 and
mpd dependencies.  It works!

If I ever learn how to asynchronously update individual components on a
lemonbar, then I will likely switch back to it.