prot-eww: Implement lynx dump as command
`prot-eww-lynx-dump' is introduced.  It depends on defcustoms
`prot-eww-lynx-dump-function' and `prot-eww-lynx-dump-dir' (see their
documentations for details).

The rational behind the *-lynx-dump-function is that, as sites contain
so much junk on their web pages, user might want clean some of that
before saving it.

The *-lynx-dump-function is called with the URL of the web page so that
user can dispatch site dependent clean up functions.