Add HerbstluftWM; reorg files; BREAKING CHANGES
I have used HerbstluftWM in the past.  In short: it feels very familiar
to someone coming from bspwm, though it is different in how it handles
monitors/workspaces.  Today I switched back to it again because I got an
ultrawide 2560x1080 display as part of a temporary exchange.  It is fun
experimenting with this setup!

The bspwm configs continue to work as intended.  But I have abstracted
away its sxhkd, polybar, and Xmodmap components, as well as a couple of
shell scripts.  Those are now part of the xorg-twm stow package, which
is shared by bswpm and herbstluftwm.

The delight script now covers herbstluftwm as well.  It underwent some
refinements, though I still plan to optimise a few things.