prot-emacs: DELETE prot-consult.el
It was really old code from the early days of Consult (I think the
recentre after jump was not there at the time).  We really don't need
any of this any more.

With regard to the wrapper functions that used orderless, I am instead
going to use 'completion-category-overrides' since now Consult has all
the metadata we need.