BREAKING delete literate config; REVISE setup
I have been postponing this for a long time.  The core idea is to reduce
the maintenance burden of my dotemacs as I already spend a lot of time
on my Emacs packages.

When I started the literate config in 2019, I did not have any packages
to maintain, so I could dedicate the requisite time to keep such a
massive document in good shape.